requirements.txt
dns.resolver==2.0.0 